requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
dj-database-url==1.0.0
Django==4.1.3
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-jazzmin==2.6.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
flake8==5.0.4
flake8-import-order==0.18.1
gunicorn==20.1.0
mccabe==0.7.0
numpy==1.23.5
phonenumbers==8.13.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
sqlparse==0.4.3
whitenoise==6.2.0
